requirements.txt
argparse==1.4.0
pytz==2022.1
attrs==20.1.0
loguru==0.6.0
more-itertools==8.5.0
packaging==20.4
psycopg2-binary==2.9.3
py==1.11.0
pyparsing==2.4.7
pytest==6.0.1
python-dotenv==0.14.0
six==1.15.0
wcwidth==0.2.5
pandas==1.3.5
clickhouse-driver==0.2.3
natsort==8.1.0
redis==4.3.3
numpy==1.21.6